United main branch layout to dev-rutas branch
